+++ src/providers/index.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router } from 'react-router-dom';
-import {Sidebar} from '../contexts/sidebar-context'
-import {FilterProvider} from '../contexts/filter-context'
+import {Sidebar} from '../contexts/sidebarContext'
+import {FilterProvider} from '../contexts/filterContext'
 
 const Providers=({children})=>{
     return <Router>
